Improve project description and language in README
Got some comments back about the language and wording use in README.

This is an attempt to improve it, adding clarifications and removing parts
already covered in other sections.

It also tries to remove the ambiguity caused by pre-compiled and binary
words used in the document.

From now on, gem-compiler generates binary gems by compiling gem's C
extensions and bundle the result into a newer gem.