Test again the version of RubyGems shipped with each version
The *only* exception for this is Ruby 1.8.7, which didn't ship
with RubyGems, however Travis and RubyInstaller have it as part
of the compiled versions.

This aims to reduce the failures in the test matrix and also the
comparison workload: test only against defaults. If you deviate
from defaults, is your fault.