Drop support for any Ruby version prior 2.0.0
Attempt to keep compatibility with Ruby 1.9.3 (or older) when these
have already reached EOL (End-Of-Life) support is too much.

Continue to support 2.0.0 until next version, but aim to drop it
*this year*.

This change also aims to *fix* the issue with AppVeyor's Ruby since
version 1.9.3 have a newer version of RubyGems that appears to fail
(and really don't want to deal with these old-versions bug reports).