Workaround Windows shortnames in temporary directories
Attempt to workaround issues around `Dir.mktmpdir` on Windows, which
might return short-filename versions of a temporary directory instead
of full-length ones like `Dir.glob`.

This might be due incorrect value of `TMP`, `TEMP` or `TMPDIR`
environment variables that is used by `Dir.mktmpdir`.

Solve this by expanding the result of `Dir.mktmpdir` to ensure both
`Dir.glob` and the directory used is consistent.

Added a test to see if this can be reproduced without luck (since is
limited to Windows environment with that specific variable).

Commit this blindly and hopefully AppVeyor will show success or failure.

Fixes #17