Remove list of personal desires from README
README included a list of unfulfilled promises that are no longer
personal desires or the direction I want this project pursue.

Lot happened since 2012 when I open-sourced this project and my
objectives for it have changed since then.

gem-compiler is now a single-purpose tool and does not aim to
become a *Jack of all trades* tool.

The removal of these items from README will help state this more
clearly.

Closes #2, closes #3 and closes #4.

[skip ci]