Lock Ruby's ABI on compiled gem
This ensures that resulting gem can only be installed on a compatible
version of Ruby, reducing the chances of an incompatible option
be exposed and the gem fail at runtime.

With this change is now the default behavior, which can be turned
off by using `--no-abi-lock` option during compilation (but is discouraged).

Closes #11