Disables test's GEM_PATH to allow `rake` work
`rake` is invoked to compile extensions as part of the tests, which
requires it to be accessible.

`GEM_PATH` is set by Gem::TestCase to use a temporary directory,
causing it not to be found and thus, triggering failures under
Windows and some Ruby/Linux combinations (with or without Bundler)

The change ensures `GEM_PATH` is unset from `ENV` on each test.