Introduce `--include-shared-dir` to compile command
Allow user to specify an additional directory to lookup for platform
specific shared libraries with the objective to bundle those inside
the final binary gem.

The directory is relative to the build location (ie. `ext` at the same
level as `lib`).

Use this only on certain scenarios where a shared library is generated
during gem installation (ie. FFI-powered one) but the artifact of such
build is not relocated to the default require directory (`lib`).

Closes #34